create lunar abstract location class
