+++ lib/zakuro/version/senmyou/stella/lunar/location.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require_relative '../../../../calculation/stella/lunar/abstract_location'
+
 require_relative '../../const/remainder'
 
 require_relative './localization'
@@ -13,7 +15,7 @@ module Zakuro
       #
       # Location 入暦
       #
-      class Location
+      class Location < Calculation::Lunar::AbstractLocation
         # @return [Cycle::LunarRemainder] 暦中日（1近点月の半分）
         HALF_ANOMALISTIC_MONTH = Const::Remainder::Lunar::HALF_ANOMALISTIC_MONTH
         # @return [Cycle::LunarRemainder] 入暦上限
@@ -21,16 +23,9 @@ module Zakuro
         # @return [Cycle::LunarRemainder] 弦
         QUARTER = Const::Remainder::Lunar::QUARTER
 
-        # @return [True] 計算済み（前回計算あり）
-        # @return [False] 未計算（初回計算）
-        attr_reader :calculated
-        # @return [Integer] 西暦年
-        attr_reader :western_year
-        # @return [True] 進（遠地点より数える）
-        # @return [False] 退（近地点より数える）
+        # @return [True] 進
+        # @return [False] 退
         attr_reader :forward
-        # @return [Cycle::LunarRemainder] 大余小余（初回：昨年天正閏余）
-        attr_reader :remainder
 
         #
         # 初期化
@@ -39,11 +34,9 @@ module Zakuro
         # @param [Integer] western_year 西暦年
         #
         def initialize(lunar_age:, western_year:)
-          @calculated = false
-          @western_year = western_year
+          super(lunar_age: lunar_age, western_year: western_year)
           # 進
           @forward = true
-          @remainder = lunar_age
         end
 
         #
@@ -63,7 +56,7 @@ module Zakuro
         # 弦の分だけ月地点を進める
         #
         def add_quarter
-          @remainder.add!(QUARTER)
+          remainder.add!(QUARTER)
         end
 
         private
@@ -73,7 +66,7 @@ module Zakuro
         #
         def first
           @remainder = Localization.first_remainder(
-            lunar_age: @remainder, western_year: @western_year
+            lunar_age: remainder, western_year: western_year
           )
           # 初回は0始まりで計算しているので、入暦上限ではなく暦中日を用いる
           decrease(limit: HALF_ANOMALISTIC_MONTH)
@@ -89,17 +82,10 @@ module Zakuro
         # @param [Cycle::LunarRemainder] limit 上限
         #
         def decrease(limit:)
-          return if @remainder < limit
+          return if remainder < limit
 
-          @remainder.sub!(HALF_ANOMALISTIC_MONTH)
-          @forward = !@forward
-        end
-
-        #
-        # 1始まりにする
-        #
-        def one_based
-          @remainder.add!(Cycle::Remainder.new(day: 1, minute: 0, second: 0))
+          remainder.sub!(HALF_ANOMALISTIC_MONTH)
+          @forward = !forward
         end
       end
     end
